+++ app/assets/stylesheets/application.scss
@@ -17,6 +17,7 @@
 @import "custom/variables";
 @import "bootstrap";
 @import "custom/theme";
+@import "font-awesome";
 @import "bookstore/general";
 @import "bookstore/utils/helpers";
 @import "variables";
