+++ app/assets/stylesheets/bookstore/_general.scss
@@ -336,4 +336,6 @@ span.error {
   border-color: red;
 }
 
-
+button.ico-cart, input.ico-cart {
+  background-color: transparent;
+}
