--- app/helpers/checkout_helper.rb
@@ -1,2 +1,0 @@
-module CheckoutHelper
-end
