--- app/helpers/order_items_helper.rb
@@ -1,2 +1,0 @@
-module OrderItemsHelper
-end
