--- app/helpers/orders_helper.rb
@@ -1,3 +1,0 @@
-module OrdersHelper
-
-end
