--- app/helpers/pages_helper.rb
@@ -1,2 +1,0 @@
-module PagesHelper
-end
