--- app/helpers/reviews_helper.rb
@@ -1,2 +1,0 @@
-module ReviewsHelper
-end
