--- app/models/order_address.rb
@@ -1,8 +1,0 @@
-class OrderAddress < ApplicationRecord
-  belongs_to :order
-  enum kind: %i[shipping billing]
-
-  include AddressValidator
-  validates_uniqueness_of :kind, scope: [:order_id]
-
-end
