+++ app/views/books/_book.html.haml
@@ -6,8 +6,10 @@
         .general-thumb-link-wrap
           = link_to book_path(book), class: "thumb-hover-link" do
             %i.fa.fa-eye.thumb-icon{'aria-hidden': "true"}
-          = link_to to_cart_path(book), class: "thumb-hover-link", method: :post do
-            %i.fa.fa-shopping-cart.thumb-icon
+          = form_for :order_item, url: order_items_path, method: :post do |f|
+            = f.hidden_field :item_id, value: book.id
+            = button_tag(type: 'submit', class: 'thumb-hover-link ico-cart') do
+              %i.fa.fa-shopping-cart.thumb-icon
   .general-thumb-info
     %p.title= book.name
     %p.font-16.in-gold-500= "#{number_to_currency(book.price, price_options)}"
