+++ app/views/layouts/partials/_footer.html.haml
@@ -8,7 +8,7 @@
           %li
             = link_to t('pages.shop'), books_path, class: "nav-link"
           %li
-            = link_to t('pages.orders'), users_orders_path, class: "nav-link"
+            = link_to t('pages.orders'), orders_users_path, class: "nav-link"
           %li
             = link_to t('pages.settings'), settings_users_path, class: "nav-link"
       .col-sm-3
@@ -27,4 +27,3 @@
           %li
             %a.nav-link{href: "#"}
               %i.fa.fa-instagram{'aria-hidden': "true"}
-%script{src: "https://use.fontawesome.com/67602d7308.js"}
