+++ app/views/layouts/partials/_top.html.haml
@@ -39,7 +39,7 @@
               %a.dropdown-toggle{href: "#", 'data-toggle': "dropdown", role: "button", 'aria-haspopup': "true", 'aria-expanded': "false"} #{t('pages.myaccount')}
               %ul.dropdown-menu
                 %li
-                  = link_to t('pages.orders'), users_orders_path
+                  = link_to t('pages.orders'), orders_users_path
                 %li
                   = link_to t('pages.settings'), settings_users_path
                 %li.divider{role: "separator"}
@@ -68,7 +68,7 @@
               %a.dropdown-toggle{href: "#", 'data-toggle': "dropdown", role: "button", 'aria-haspopup': "true", 'aria-expanded': "false"} #{t('pages.myaccount')}
               %ul.dropdown-menu
                 %li
-                  = link_to t('pages.orders'), users_orders_path
+                  = link_to t('pages.orders'), orders_users_path
                 %li
                   = link_to t('pages.settings'), settings_users_path
                 %li.divider{role: "separator"}
