+++ config/initializers/locale.rb
@@ -1,3 +1,3 @@
 I18n.load_path += Dir[Rails.root.join('lib', 'locale', '*.{rb,yml}')]
-I18n.available_locales = [:en, :ru]
+I18n.available_locales = [:en]
 I18n.default_locale = :en
