+++ spec/controllers/order_items_controller_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe OrderItemsController, type: :controller do
 
       it 'add item to order' do
         _quantity = 1
-        post :create, params: { item_id: book.id, quantity: _quantity }
+        post :create, params: { order_item: { item_id: book.id, quantity: _quantity } }
         expect(flash[:notice]).to match I18n.t('order_item.create_success')
         order.reload
         expect(order.order_items[0].item_id).to eq book.id
@@ -23,13 +23,13 @@ RSpec.describe OrderItemsController, type: :controller do
       it 'found the same item and only change count' do
         _quantity = 1
         _added_val = 3
-        post :create, params: { item_id: book.id, quantity: _quantity }
+        post :create, params: { order_item: { item_id: book.id, quantity: _quantity } }
         order.reload
 
         expect(order.order_items[0].item_id).to eq book.id
         expect(order.order_items[0].quantity).to eq _quantity
 
-        post :create, params: { item_id: book.id, quantity: _added_val }
+        post :create, params: { order_item: { item_id: book.id, quantity: _added_val } }
         order.reload
         expect(order.order_items[0].item_id).to eq book.id
         expect(order.order_items[0].quantity).to eq (_quantity + _added_val)
@@ -38,14 +38,14 @@ RSpec.describe OrderItemsController, type: :controller do
       it 'should not add item to order' do
         _quantity = 1
         rand_id = "-1232352"
-        post :create, params: { item_id: rand_id, quantity: _quantity }
+        post :create, params: { order_item: { item_id: rand_id, quantity: _quantity } }
         expect(flash[:notice]).to match I18n.t('order_item.create_fail')
         order.reload
         expect(order.order_items.size).to be_zero
       end
     end
 
-    context 'POST #update' do
+    context 'PATCH #update' do
       before(:each) do
         sign_in user
         @order_item = create(:order_item, book: book)
@@ -55,20 +55,20 @@ RSpec.describe OrderItemsController, type: :controller do
 
       it 'change quantity item' do
         _new_quantity = @order_item.quantity + 5
-        post :update, params: { id: @order_item.id, quantity: _new_quantity }
+        patch :update, params: { id: @order_item.id, order_item: { quantity: _new_quantity } }
         order.reload
         expect(order.order_items[0].quantity).to eq(_new_quantity)
       end
 
       it 'should not change quantity' do
         rand_id = "-1232352"
-        post :update, params: { id: rand_id, quantity: 1 }
+        patch :update, params: { id: rand_id, order_item: { quantity: 1 } }
         expect(flash[:alert]).to match I18n.t('order_item.update_fail')
       end
 
       it 'set quantity to 1 if new value is not valid' do
         invalid_value = -7
-        post :update, params: { id: @order_item.id, quantity: invalid_value }
+        patch :update, params: { id: @order_item.id, order_item: { quantity: invalid_value } }
         order.reload
         expect(order.order_items[0].quantity).to eq 1
       end
