--- spec/helpers/checkout_helper_spec.rb
@@ -1,15 +1,0 @@
-require 'rails_helper'
-
-# Specs in this file have access to a helper object that includes
-# the CheckoutHelper. For example:
-#
-# describe CheckoutHelper do
-#   describe "string concat" do
-#     it "concats two strings with spaces" do
-#       expect(helper.concat_strings("this","that")).to eq("this that")
-#     end
-#   end
-# end
-RSpec.describe CheckoutHelper, type: :helper do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
